+++ src/v/archival/ntp_archiver_service.cc
@@ -20,8 +20,6 @@
 #include "cluster/partition_manager.h"
 #include "config/configuration.h"
 #include "model/metadata.h"
-#include "s3/error.h"
-#include "s3/s3_client.h"
 #include "storage/disk_log_impl.h"
 #include "storage/fs_utils.h"
 #include "storage/parser.h"
