+++ src/v/archival/service.cc
@@ -27,8 +27,6 @@
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/namespace.h"
-#include "s3/error.h"
-#include "s3/s3_client.h"
 #include "storage/disk_log_impl.h"
 #include "storage/fs_utils.h"
 #include "storage/log.h"
