+++ src/v/cloud_storage/partition_recovery_manager.h
@@ -14,7 +14,6 @@
 #include "cloud_storage/remote.h"
 #include "model/metadata.h"
 #include "model/record.h"
-#include "s3/s3_client.h"
 #include "storage/ntp_config.h"
 #include "utils/retry_chain_node.h"
