+++ src/v/cloud_storage/remote.cc
@@ -16,8 +16,6 @@
 #include "cloud_storage/types.h"
 #include "model/metadata.h"
 #include "net/connection.h"
-#include "s3/configuration.h"
-#include "s3/s3_client.h"
 #include "ssx/sformat.h"
 #include "utils/intrusive_list_helpers.h"
 #include "utils/retry_chain_node.h"
