+++ src/v/cloud_storage/tests/remote_partition_test.cc
@@ -25,8 +25,7 @@
 #include "model/record.h"
 #include "model/record_batch_types.h"
 #include "model/timeout_clock.h"
-#include "s3/configuration.h"
-#include "s3/s3_client.h"
+#include "s3/types.h"
 #include "seastarx.h"
 #include "storage/log.h"
 #include "storage/log_manager.h"
