+++ src/v/cloud_storage/tests/remote_segment_test.cc
@@ -18,9 +18,9 @@
 #include "cloud_storage/tests/cloud_storage_fixture.h"
 #include "cloud_storage/tests/common_def.h"
 #include "cloud_storage/types.h"
+#include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/timeout_clock.h"
-#include "s3/s3_client.h"
 #include "seastarx.h"
 #include "storage/log.h"
 #include "storage/log_manager.h"
