+++ src/v/cloud_storage/tests/remote_test.cc
@@ -18,8 +18,6 @@
 #include "cloud_storage/tests/s3_imposter.h"
 #include "cloud_storage/types.h"
 #include "model/metadata.h"
-#include "s3/client_probe.h"
-#include "s3/s3_client.h"
 #include "seastarx.h"
 #include "storage/directories.h"
 #include "test_utils/async.h"
